backend/requirements.txt
# ==================== 后端依赖 ====================
# FastAPI 框架
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.0.0

# 图像处理
Pillow>=10.0.0
opencv-python-headless>=4.9.0
numpy>=1.26.0,<2
scipy>=1.12.0

# 机器学习
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.4.0

# DINOv3 依赖
ftfy
omegaconf
regex

# SAM3 依赖
timm>=1.0.17
huggingface_hub
tqdm
typing_extensions

# YOLO-World 依赖 (需要 OpenMIM 安装 mmcv)
# pip install openmim && mim install mmcv==2.0.0
mmengine>=0.10.3
mmdet>=3.0.0
transformers>=4.36.0
